Move journal shaper and beautifier into their own files
